Move `extend` to `util/core`
