+++ lib/sinon.js
@@ -43,7 +43,7 @@ function exposeEventTarget(target, eventTarget) {
 exposeCoreUtils(exports, require("./sinon/util/core/index"));
 
 // TODO move these two into util/core.
-exports.extend = deprecated.wrap(require("./sinon/extend"), deprecated.defaultMsg("extend"));
+exports.extend = deprecated.wrap(require("./sinon/util/core/extend"), deprecated.defaultMsg("extend"));
 exports.typeOf = deprecated.wrap(require("./sinon/util/core/typeOf"), deprecated.defaultMsg("typeOf"));
 
 exports.assert = require("./sinon/assert");
