+++ lib/sinon/behavior.js
@@ -9,7 +9,7 @@
  */
 "use strict";
 
-var extend = require("./extend");
+var extend = require("./util/core/extend");
 var functionName = require("./util/core/function-name");
 var valueToString = require("./util/core/value-to-string");
