+++ lib/sinon/mock-expectation.js
@@ -11,7 +11,7 @@
 var spyInvoke = require("./spy").invoke;
 var spyCallToString = require("./call").toString;
 var timesInWords = require("./util/core/times-in-words");
-var extend = require("./extend");
+var extend = require("./util/core/extend");
 var match = require("./match");
 var stub = require("./stub");
 var assert = require("./assert");
