+++ lib/sinon/mock.js
@@ -10,7 +10,7 @@
 
 var mockExpectation = require("./mock-expectation");
 var spyCallToString = require("./call").toString;
-var extend = require("./extend");
+var extend = require("./util/core/extend");
 var match = require("./match");
 var deepEqual = require("./util/core/deep-equal").use(match);
 var wrapMethod = require("./util/core/wrap-method");
