+++ lib/sinon/sandbox.js
@@ -9,7 +9,7 @@
  */
 "use strict";
 
-var extend = require("./extend");
+var extend = require("./util/core/extend");
 var createInstance = require("./util/core/create");
 var sinonCollection = require("./collection");
 var sinonMatch = require("./match");
