+++ lib/sinon/spy.js
@@ -9,7 +9,7 @@
 "use strict";
 require("colors");
 
-var extend = require("./extend");
+var extend = require("./util/core/extend");
 var functionName = require("./util/core/function-name");
 var functionToString = require("./util/core/function-to-string");
 var getPropertyDescriptor = require("./util/core/get-property-descriptor");
