+++ lib/sinon/stub.js
@@ -10,7 +10,7 @@
 
 var behavior = require("./behavior");
 var spy = require("./spy");
-var extend = require("./extend");
+var extend = require("./util/core/extend");
 var walk = require("./util/core/walk");
 var objectKeys = require("./util/core/object-keys");
 var getPropertyDescriptor = require("./util/core/get-property-descriptor");
