+++ lib/sinon/util/core/extend.js
