+++ lib/sinon/util/fake_xdomain_request.js
@@ -5,7 +5,7 @@
 "use strict";
 
 var event = require("./event");
-var extend = require("../extend");
+var extend = require("./core/extend");
 var configureLogError = require("./core/log_error");
 
 var xdr = { XDomainRequest: global.XDomainRequest };
