+++ lib/sinon/util/fake_xml_http_request.js
@@ -12,7 +12,7 @@ var TextEncoder = require("text-encoding").TextEncoder;
 
 var configureLogError = require("./core/log_error");
 var sinonEvent = require("./event");
-var extend = require("../extend");
+var extend = require("./core/extend");
 
 function getWorkingXHR(globalScope) {
     var supportsXHR = typeof globalScope.XMLHttpRequest !== "undefined";
