+++ test/util/fake-xml-http-request-test.js
@@ -3,7 +3,7 @@
 var referee = require("referee");
 var sinonStub = require("../../lib/sinon/stub");
 var sinonSpy = require("../../lib/sinon/spy");
-var sinonExtend = require("../../lib/sinon/extend");
+var sinonExtend = require("../../lib/sinon/util/core/extend");
 var sinonSandbox = require("../../lib/sinon/sandbox");
 var sinonFakeXhr = require("../../lib/sinon/util/fake_xml_http_request");
 var sinon = require("../../lib/sinon");
